+++ src/DoctrineMongoODMModule/Module.php
@@ -35,8 +35,6 @@ class Module
     {
         $commands = [
             new Command\QueryCommand(),
-            new Command\GenerateDocumentsCommand(),
-            new Command\GenerateRepositoriesCommand(),
             new Command\GenerateProxiesCommand(),
             new Command\GenerateHydratorsCommand(),
             new Command\GeneratePersistentCollectionsCommand(),