+++ tests/DoctrineMongoODMModuleTest/Doctrine/AnnotationTest.php
@@ -10,6 +10,7 @@ class AnnotationTest extends AbstractTest
 {
     public function testAnnotation() : void
     {
+        self::markTestIncomplete();
         $documentManager = $this->getDocumentManager();
         $metadata        = $documentManager->getClassMetadata('DoctrineMongoODMModuleTest\Assets\Document\Annotation');
     }